+++ dvb-si/src/descriptors/mod.rs
@@ -79,6 +79,7 @@ pub mod multiplex_buffer_utilization;
 pub mod muxcode;
 pub mod mvc_extension;
 pub mod network_name;
+pub mod nordig;
 pub mod nvod_reference;
 pub mod parental_rating;
 pub mod partial_transport_stream;